web-bundle/subresource-loading/link-relative-url-in-web-bundle.https.tentative.html

<!DOCTYPE html>
<title>Relative Url in web bundle</title>
<link rel="help" href="https://github.com/WICG/webpackage/blob/main/explainers/subresource-loading.md" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/test-helpers.js"></script>

<body>
  <link
  rel="webbundle"
  href="../resources/wbn/relative-url.wbn"
  resources="
    ../resources/wbn/relative-url-file.js
    ../resources/wbn/relative-url/start-with-double-slash.js
    ../resources/wbn/relative-url/start-with-slash.js
    ../resources/wbn/relative-url/subdirectory-path.js
    " />
  <script>
    const loaded_scripts = [];
    function scriptLoaded(file) {
      loaded_scripts.push(file);
    }
  </script>
  <script src="../resources/wbn/relative-url-file.js"></script>
  <script src="../resources/wbn/relative-url/start-with-double-slash.js"></script>
  <script src="../resources/wbn/relative-url/start-with-slash.js"></script>
  <script src="../resources/wbn/relative-url/subdirectory-path.js"></script>
<body>
  <script>
    promise_test(async (t) => {
      assert_array_equals(
        loaded_scripts,
        [
          'relative-url-file.js',
          'start-with-double-slash.js',
          'start-with-slash.js',
          'subdirectory-path.js',
        ]);
    },
    'Relative Url in web bundle.');
  </script>
</body>
